Update skeleton.c
Went back to earlier version of code with the findMeanVector modifies *meanVector in place instead of outputting pointer. Realized the E step doesn't require calculating mean vector for each component. The "mean vector" in the text is the center, which is initialized and then updated in the M step.

Wrote the E step and parallelized row by row using omp parallel for. Needed to write support functions find_determinant, gaussJordan (for the inverse) and pdf.